+++ src/SpeakEasy.Specifications/FileUploadBodySpecification.cs
@@ -1,5 +1,6 @@
 ﻿using Machine.Fakes;
 using Machine.Specifications;
+using SpeakEasy.Bodies;
 
 namespace SpeakEasy.Specifications
 {
