+++ src/SpeakEasy.Specifications/SpeakEasy.Specifications.csproj
@@ -74,6 +74,7 @@
     </Compile>
     <Compile Include="Authenticators\BasicAuthenticatorSpecification.cs" />
     <Compile Include="Authenticators\WindowsAuthenticatorSpecification.cs" />
+    <Compile Include="Bodies\MultipartMimeDocumentBodySpecification.cs" />
     <Compile Include="MultipleValuesArrayFormatterSpecification.cs" />
     <Compile Include="PersistentCookieStrategySpecification.cs" />
     <Compile Include="TransientCookieStrategySpecification.cs" />
@@ -84,7 +85,6 @@
     <Compile Include="HeaderSpecification.cs" />
     <Compile Include="HeadRequestSpecification.cs" />
     <Compile Include="HttpRequestSpecification.cs" />
-    <Compile Include="MultipartMimeDocumentBodySpecification.cs" />
     <Compile Include="OptionsRequestSpecification.cs" />
     <Compile Include="ParameterSpecification.cs" />
     <Compile Include="ParsedHeaderSpecification.cs" />
