+++ src/SpeakEasy/PostRequestBody.cs
@@ -1,4 +1,5 @@
 using System.Text;
+using SpeakEasy.Bodies;
 
 namespace SpeakEasy
 {
