+++ src/SpeakEasy/SpeakEasy.csproj
@@ -46,6 +46,10 @@
     <Compile Include="Authenticators\BasicAuthenticator.cs" />
     <Compile Include="Authenticators\NullAuthenticator.cs" />
     <Compile Include="Authenticators\WindowsAuthenticator.cs" />
+    <Compile Include="Bodies\MultipartMimeDocumentBody.cs" />
+    <Compile Include="Bodies\NullSerializableBody.cs" />
+    <Compile Include="Bodies\SerializableByteArray.cs" />
+    <Compile Include="Bodies\StreamableSerializableBody.cs" />
     <Compile Include="CommaSeparatedArrayFormatter.cs" />
     <Compile Include="ConfigurationException.cs" />
     <Compile Include="Cookie.cs" />
@@ -56,7 +60,6 @@
     <Compile Include="MultipleValuesArrayFormatter.cs" />
     <Compile Include="PersistentCookieStrategy.cs" />
     <Compile Include="SimpleJson.cs" />
-    <Compile Include="StreamableSerializableBody.cs" />
     <Compile Include="TransientCookieStrategy.cs" />
     <Compile Include="DefaultNamingConvention.cs" />
     <Compile Include="DeleteRequest.cs" />
@@ -79,19 +82,16 @@
     <Compile Include="IRequestBody.cs" />
     <Compile Include="IUserAgent.cs" />
     <Compile Include="NullRequestBody.cs" />
-    <Compile Include="NullSerializableBody.cs" />
     <Compile Include="PostRequestBody.cs" />
     <Compile Include="IResourceMerger.cs" />
     <Compile Include="ISerializableBody.cs" />
     <Compile Include="Loggers\ConsoleLogger.cs" />
     <Compile Include="Loggers\NullLogger.cs" />
-    <Compile Include="MultipartMimeDocumentBody.cs" />
     <Compile Include="ObjectRequestBody.cs" />
     <Compile Include="ParsedHeaderValue.cs" />
     <Compile Include="ParsedHeaderParameter.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ResourceMerger.cs" />
-    <Compile Include="SerializableByteArray.cs" />
     <Compile Include="SerializerException.cs" />
     <Compile Include="Serializers\NullSerializer.cs" />
     <Compile Include="FileUpload.cs" />
@@ -135,5 +135,6 @@
   <ItemGroup>
     <None Include="packages.config" />
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(SolutionDir)\.nuget\nuget.targets" />
 </Project>
